api_ruc_python/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "API para consultar información de RUC usando IA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.105.0",
    "uvicorn>=0.24.0",
    "langchain-openai>=0.3.11",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api_ruc_python"]
